refactor(test): Changes file structure a bit
