+++ test/on-chain-data/hotels/wt-index.spec.js
@@ -1,18 +1,17 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
-import WTIndexDataProvider from '../../../../src/on-chain-data/airlines/wt-index';
-import { AirlineDataModel } from '../../../../src/on-chain-data/';
-import testedDataModel from '../../../utils/data-airline-model-definition';
+import WTIndexDataProvider from '../../../src/on-chain-data/hotels/wt-index';
+import { HotelDataModel } from '../../../src/on-chain-data/';
+import testedDataModel from '../../utils/data-hotel-model-definition';
+import { WTLibsError } from '../../../src/errors';
+import { SmartContractInstantiationError, HotelNotInstantiableError,
+  HotelNotFoundError, RemoteDataReadError } from '../../../src/on-chain-data/errors';
 
-import { WTLibsError } from '../../../../src/errors';
-import { SmartContractInstantiationError, AirlineNotInstantiableError, AirlineNotFoundError,
-  RemoteDataReadError } from '../../../../src/on-chain-data/errors';
-
-describe('WTLibs.on-chain-data.airlines.WTAirlineIndex', () => {
+describe('WTLibs.on-chain-data.hotels.WTHotelIndex', () => {
   let dataModel, indexDataProvider;
 
   beforeEach(async function () {
-    dataModel = AirlineDataModel.createInstance(testedDataModel.withDataSource().dataModelOptions);
+    dataModel = HotelDataModel.createInstance(testedDataModel.withDataSource().dataModelOptions);
     indexDataProvider = dataModel.getWindingTreeIndex(testedDataModel.indexAddress);
   });
 
@@ -22,40 +21,40 @@ describe('WTLibs.on-chain-data.airlines.WTAirlineIndex', () => {
       await customIndexDataProvider._getDeployedIndex();
       throw new Error('should not have been called');
     } catch (e) {
-      assert.match(e.message, /cannot get airlineIndex instance/i);
+      assert.match(e.message, /cannot get hotelIndex instance/i);
       assert.instanceOf(e, SmartContractInstantiationError);
     }
   });
 
-  describe('getAirline', () => {
+  describe('getHotel', () => {
     it('should throw if address is malformed', async () => {
       try {
-        await indexDataProvider.getAirline('random-address');
+        await indexDataProvider.getHotel('random-address');
         throw new Error('should not have been called');
       } catch (e) {
-        assert.match(e.message, /cannot find airline/i);
+        assert.match(e.message, /cannot find hotel/i);
         assert.instanceOf(e, WTLibsError);
       }
     });
 
-    it('should throw if no airline exists on that address', async () => {
+    it('should throw if no hotel exists on that address', async () => {
       try {
-        await indexDataProvider.getAirline('0x96eA4BbF71FEa3c9411C1Cefc555E9d7189695fA');
+        await indexDataProvider.getHotel('0x96eA4BbF71FEa3c9411C1Cefc555E9d7189695fA');
         throw new Error('should not have been called');
       } catch (e) {
-        assert.match(e.message, /cannot find airline/i);
-        assert.instanceOf(e, AirlineNotFoundError);
+        assert.match(e.message, /cannot find hotel/i);
+        assert.instanceOf(e, HotelNotFoundError);
       }
     });
 
-    it('should throw if airline contract cannot be instantiated', async () => {
+    it('should throw if hotel contract cannot be instantiated', async () => {
       try {
         sinon.stub(indexDataProvider, '_createRecordInstanceFactory').rejects();
-        await indexDataProvider.getAirline('0x820410b0E5c06147f1a894247C46Ea936D8A4Eb8');
+        await indexDataProvider.getHotel('0xbf18b616ac81830dd0c5d4b771f22fd8144fe769');
         throw new Error('should not have been called');
       } catch (e) {
-        assert.match(e.message, /cannot find airline/i);
-        assert.instanceOf(e, AirlineNotInstantiableError);
+        assert.match(e.message, /cannot find hotel/i);
+        assert.instanceOf(e, HotelNotInstantiableError);
       } finally {
         indexDataProvider._createRecordInstanceFactory.restore();
       }
@@ -64,65 +63,65 @@ describe('WTLibs.on-chain-data.airlines.WTAirlineIndex', () => {
     it('should throw if accessing off-chain data without resolved on-chain pointer and on-chain pointer cannot be downloaded', async () => {
       // pre-heat the contract so we can stub it later
       await indexDataProvider._getDeployedIndex();
-      sinon.stub(indexDataProvider.deployedIndex.methods, 'airlinesIndex').returns({
+      sinon.stub(indexDataProvider.deployedIndex.methods, 'hotelsIndex').returns({
         call: sinon.stub().resolves('7'),
       });
-      // There is not a valid airline on this address
+      // There is not a valid hotel on this address
       const address = '0x994afd347b160be3973b41f0a144819496d175e9';
-      const airline = await indexDataProvider.getAirline(address);
+      const hotel = await indexDataProvider.getHotel(address);
 
       try {
-        await airline.dataIndex;
+        await hotel.dataIndex;
         throw new Error('should not have been called');
       } catch (e) {
         assert.match(e.message, /cannot sync remote data/i);
         assert.instanceOf(e, RemoteDataReadError);
       } finally {
-        indexDataProvider.deployedIndex.methods.airlinesIndex.restore();
+        indexDataProvider.deployedIndex.methods.hotelsIndex.restore();
       }
     });
   });
 
-  describe('getAllAirlines', () => {
+  describe('getAllHotels', () => {
     it('should call getAllRecords', async () => {
       const callStub = sinon.spy(indexDataProvider, 'getAllRecords');
-      await indexDataProvider.getAllAirlines();
+      await indexDataProvider.getAllHotels();
       assert.equal(callStub.callCount, 1);
       indexDataProvider.getAllRecords.restore();
     });
   });
 
-  describe('transferAirlineOwnership', () => {
+  describe('transferHotelOwnership', () => {
     it('should call transferRecordOwnership', async () => {
       const callStub = sinon.stub(indexDataProvider, 'transferRecordOwnership').resolves(true);
-      await indexDataProvider.transferAirlineOwnership({}, '0x820410b0E5c06147f1a894247C46Ea936D8A4Eb8');
+      await indexDataProvider.transferHotelOwnership({}, '0x820410b0E5c06147f1a894247C46Ea936D8A4Eb8');
       assert.equal(callStub.callCount, 1);
       indexDataProvider.transferRecordOwnership.restore();
     });
   });
 
-  describe('addAirline', () => {
+  describe('addHotel', () => {
     it('should call addRecord', async () => {
       const callStub = sinon.stub(indexDataProvider, 'addRecord').resolves(true);
-      await indexDataProvider.addAirline({});
+      await indexDataProvider.addHotel({});
       assert.equal(callStub.callCount, 1);
       indexDataProvider.addRecord.restore();
     });
   });
 
-  describe('updateAirline', () => {
+  describe('updateHotel', () => {
     it('should call updateRecord', async () => {
       const callStub = sinon.stub(indexDataProvider, 'updateRecord').resolves(true);
-      await indexDataProvider.updateAirline({});
+      await indexDataProvider.updateHotel({});
       assert.equal(callStub.callCount, 1);
       indexDataProvider.updateRecord.restore();
     });
   });
 
-  describe('removeAirline', () => {
+  describe('removeHotel', () => {
     it('should call removeRecord', async () => {
       const callStub = sinon.stub(indexDataProvider, 'removeRecord').resolves(true);
-      await indexDataProvider.removeAirline('0x820410b0E5c06147f1a894247C46Ea936D8A4Eb8');
+      await indexDataProvider.removeHotel('0x820410b0E5c06147f1a894247C46Ea936D8A4Eb8');
       assert.equal(callStub.callCount, 1);
       indexDataProvider.removeRecord.restore();
     });
