+++ test/on-chain-data/storage-pointer.spec.js
@@ -1,10 +1,10 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
 import InMemoryAdapter from '@windingtree/off-chain-adapter-in-memory';
-import StoragePointer from '../../../src/on-chain-data/storage-pointer';
-import { StoragePointerError } from '../../../src/on-chain-data/errors';
-import { OffChainDataClient } from '../../../src/off-chain-data-client';
-import { OffChainDataRuntimeError } from '../../../src/off-chain-data-client/errors';
+import StoragePointer from '../../src/on-chain-data/storage-pointer';
+import { StoragePointerError } from '../../src/on-chain-data/errors';
+import { OffChainDataClient } from '../../src/off-chain-data-client';
+import { OffChainDataRuntimeError } from '../../src/off-chain-data-client/errors';
 
 describe('WTLibs.StoragePointer', () => {
   beforeEach(() => {
