+++ test/on-chain-data/wt-index/index.spec.js
@@ -1,13 +1,12 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
-import AbstractWTIndex from '../../../../src/on-chain-data/wt-index';
+import AbstractWTIndex from '../../../src/on-chain-data/wt-index';
 
-import { HotelDataModel } from '../../../../src/on-chain-data/';
-import testedDataModel from '../../../utils/data-hotel-model-definition';
-
-import { WTLibsError } from '../../../../src/errors';
+import { HotelDataModel } from '../../../src/on-chain-data/';
+import testedDataModel from '../../utils/data-hotel-model-definition';
+import { WTLibsError } from '../../../src/errors';
 import { RecordNotFoundError, RecordNotInstantiableError,
-  InputDataError } from '../../../../src/on-chain-data/errors';
+  InputDataError } from '../../../src/on-chain-data/errors';
 
 describe('WTLibs.on-chain-data.AbstractWTIndex', () => {
   let dataModel;
