+++ test/wallet/index.spec.js
